pyproject.toml
[project]
name = "fast-api-with-docker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "docling>=2.70.0",
    "feedparser>=6.0.12",
    "google-genai>=1.61.0",
    "google-generativeai>=0.8.6",
    "markdown>=3.10.1",
    "markdownify>=1.2.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "youtube-transcript-api>=1.2.4",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
